+++ .ci/after_failure.sh
@@ -2,12 +2,9 @@
 
 set -ex
 
-cd "${HOME}/workspace"
-
-if [ "${TRAVIS_OS_NAME}" = "linux" ]; then
-  cat ./build/aikido/Testing/Temporary/LastTest.log
-  cat ./build/aikido/Testing/Temporary/LastTestsFailed.log
-elif [ "${TRAVIS_OS_NAME}" = "osx"   ]; then
-  cat Testing/Temporary/LastTest.log
-  cat Testing/Temporary/LastTestsFailed.log
+if [ "${USE_CATKIN}" = "ON" ]; then
+  cd "${HOME}/workspace/build/aikido/"
 fi
+
+cat Testing/Temporary/LastTest.log
+cat Testing/Temporary/LastTestsFailed.log